Migrate `RecentPatientItem` to use `ItemAdapter`
